+++ hacker-stories/src/App.jsx
@@ -1,50 +1,31 @@
-import * as React from 'react';
+  import * as React from 'react';
+  import List from "./List";
+  import Search from './Search';
+ 
+ 
 
-const list = [
-  {
-    title: 'React' ,
-    url: 'https://reactjs.org/' ,
-    author: 'Jordan Walke' ,
-    num_comments: 3,
-    points: 4 ,
-    objectID: 0,
-  },
-  {
-    title: 'Redux' ,
-    url: 'https://reduxjs.org/' ,
-    author: 'Dan Abramov, Andrew Clark' ,
-    num_comments: 2,
-    points: 5,
-    objectID: 1,
-  }
-]
+ 
 
-function App() {
-  return (
+  function App() {
+    return (
     <div>
-      <h1>My Hacker Stories</h1>
-      <label htmlFor="search">Search</label>
-      <input is="search" type="text" />
+        <h1>My Hacker Stories</h1>
+        
 
-      <hr />
+        <Search />
 
-      <ul>
-        {list.map(function(item) {
-          return (
-            <li key={item.objectID}>
-              <span>
-                <a href={item.url}>{item.title}</a>
-              </span>
-              <span>{item.author}</span>
-              <span>{item.num_comments}</span>
-              <span>{item.points}</span>
-            </li>
-          )
-         
-        })}
-      </ul>
-    </div>
-  );
-}
+        <hr />
 
-export default App;
+        <List />
+        <List />
+
+      
+  </div>
+    );
+  }
+
+  
+  
+
+
+  export default App;
