+++ hacker-stories/src/Search.jsx
@@ -1,0 +1,14 @@
+
+
+function Search() {
+    return (
+      <div>
+      <label htmlFor="search">Search</label>
+        <input id="search" type="text" />
+      </div>
+    )
+
+  }
+
+
+  export default Search;
